+++ src/schema/edtr-io-types.ts
@@ -7,7 +7,7 @@ import { GeogebraPluginState } from '@/serlo-editor/plugins/geogebra'
 import { HighlightPluginState } from '@/serlo-editor/plugins/highlight'
 import { ImagePluginState } from '@/serlo-editor/plugins/image'
 import { InjectionPluginState } from '@/serlo-editor/plugins/injection'
-import { MultimediaExplanationPluginState } from '@/serlo-editor/plugins/multimedia-explanation'
+import { MultimediaExplanationPluginState } from '@/serlo-editor/plugins/multimedia'
 import { PageLayoutPluginState } from '@/serlo-editor/plugins/page-layout'
 import { PageTeamPluginState } from '@/serlo-editor/plugins/page-team'
 import { RowsPluginState } from '@/serlo-editor/plugins/rows'
