+++ src/serlo-editor-integration/plugins.tsx
@@ -28,7 +28,7 @@ import { createHighlightPlugin } from '@/serlo-editor/plugins/highlight'
 import { imagePlugin } from '@/serlo-editor/plugins/image/image-with-serlo-config'
 import { injectionPlugin } from '@/serlo-editor/plugins/injection'
 import { createInputExercisePlugin } from '@/serlo-editor/plugins/input-exercise'
-import { createMultimediaExplanationPlugin } from '@/serlo-editor/plugins/multimedia-explanation'
+import { createMultimediaPlugin } from '@/serlo-editor/plugins/multimedia'
 import { pageLayoutPlugin } from '@/serlo-editor/plugins/page-layout'
 import { pagePartnersPlugin } from '@/serlo-editor/plugins/page-partners'
 import { pageTeamPlugin } from '@/serlo-editor/plugins/page-team'
@@ -97,7 +97,7 @@ export function createPlugins({
     { type: 'image', plugin: imagePlugin, visible: true, icon: <IconImage /> },
     {
       type: 'multimedia',
-      plugin: createMultimediaExplanationPlugin(),
+      plugin: createMultimediaPlugin(),
       visible: true,
       icon: <IconMultimedia />,
     },
@@ -179,7 +179,7 @@ export function createPlugins({
     { type: 'article', plugin: articlePlugin },
     {
       type: 'articleIntroduction',
-      plugin: createMultimediaExplanationPlugin({
+      plugin: createMultimediaPlugin({
         explanation: {
           plugin: 'text',
           config: {
