+++ src/serlo-editor/plugins/multimedia/resizable.tsx
